One more long -> int fix
git-svn-id: http://svn.boost.org/svn/boost/trunk@28016 b8fc166d-592f-0410-95f2-cb63ce0dd405

+++ boost/boost/detail/sp_counted_base_gcc_x86.hpp
@@ -162,7 +162,7 @@ public:
 
     long use_count() const // nothrow
     {
-        return static_cast<long const volatile &>( use_count_ );
+        return static_cast<int const volatile &>( use_count_ );
     }
 };
